adding palette and testing it in home
