+++ package.json
@@ -33,7 +33,9 @@
     "file-loader": "4.0.0",
     "html-webpack-plugin": "3.2.0",
     "mini-css-extract-plugin": "^0.7.0",
+    "node-sass": "4.12.0",
     "optimize-css-assets-webpack-plugin": "5.0.1",
+    "sass-loader": "7.1.0",
     "style-loader": "0.23.1",
     "terser-webpack-plugin": "1.3.0",
     "url-loader": "2.0.0",
