+++ src/containers/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
-import Home from '../pages/Home.js';
+import Home from '../containers/Home.js';
 
 const App = () => {
   console.log('TESTING REACT');
