+++ src/containers/Home.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import '../styles/scss/main.scss';
+
+const Home = () => {
+  return (
+    <>
+     <h1>HOLA OOZMAKAPPA.JS aqui va el logo bebé</h1>
+      <p className="sass">esto es sass</p>
+    </>
+  )
+}
+
+export default Home;
