--- src/pages/Home.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Home = () => {
-  console.log('HELLO HOME :D')
-  return (
-    <h1>HOLA OOZMAKAPPA.JS :D</h1>
-  )
-}
-
-export default Home;
