+++ src/styles/scss/_palette.scss
@@ -1,0 +1,28 @@
+$palette: (
+    primary: (
+      blue: #5BDBC3,
+      purple:#594592,
+    ),
+    lightTheme: (
+        background: #F2F6F5,
+        content:#2B2B28,
+        lighter: #48525C
+    ),
+    darkTheme: (
+        background: #222831,
+        content: #EBEBE3,
+    ),
+    greyColors: (
+      moon: #999999,
+      snow: #F4F4F4,
+      nickel: #666666,
+      coal:#333333,
+      silver: #CCCCCC,
+      black: #000000,
+    ),
+    alert: (
+      failure:#A50907,
+      warning: #F9C805,
+      success: #2DB428,
+    )
+);
