+++ src/styles/scss/main.scss
@@ -1,0 +1,13 @@
+@import 'palette';
+
+body {
+  background-color: map-get(map-get($palette, darkTheme), background);
+  display: flex;
+  width: 100%;
+  height: 100vh;
+  justify-content: center;
+  align-items: center;
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  color: map-get(map-get($palette, primary), purple);
+  font-size: 1em;
+}
