+++ webpack.dev.config.js
@@ -26,10 +26,23 @@ module.exports = {
         exclude: /node_modules/,
       },
       {
+        test: /\.scss$/,
+        use: [
+          'style-loader',
+          'css-loader',
+          'sass-loader',
+        ]
+      },
+      {
         test: /\.css$/,
         use: [
           'style-loader',
-          'css-loader'
+         {
+          loader: 'css-loader',
+          options: {
+            importLoaders: 1
+          }
+         },
         ]
       },
       {
@@ -50,3 +63,5 @@ module.exports = {
     }),
   ],
 }
+
+
